chore: remove ts3.6 support for node
