--- types/node/ts3.6/index.d.ts
@@ -1,7 +1,0 @@
-// NOTE: These definitions support NodeJS and TypeScript 3.6.
-// This is required to enable typing assert in ts3.7 without causing errors
-// Typically type modifications should be made in base.d.ts instead of here
-
-/// <reference path="base.d.ts" />
-
-/// <reference path="assert.d.ts" />
