--- types/node/ts3.6/tsconfig.json
@@ -1,24 +1,0 @@
-{
-    "files": [
-        "index.d.ts",
-        "node-tests.ts"
-    ],
-    "compilerOptions": {
-        "module": "commonjs",
-        "target": "esnext",
-        "lib": [
-            "dom"
-        ],
-        "noImplicitAny": true,
-        "noImplicitThis": true,
-        "strictNullChecks": true,
-        "strictFunctionTypes": true,
-        "baseUrl": "../../",
-        "typeRoots": [
-            "../../"
-        ],
-        "types": [],
-        "noEmit": true,
-        "forceConsistentCasingInFileNames": true
-    }
-}
