--- types/node/v12/package.json
@@ -1,11 +1,0 @@
-{
-    "private": true,
-    "types": "index",
-    "typesVersions": {
-        "<=3.6": {
-            "*": [
-                "ts3.6/*"
-            ]
-        }
-    }
-}
