--- types/node/v14/ts3.6/tsconfig.json
@@ -1,30 +1,0 @@
-{
-    "files": [
-        "index.d.ts",
-        "node-tests.ts"
-    ],
-    "compilerOptions": {
-        "module": "commonjs",
-        "target": "esnext",
-        "lib": [
-            "es6",
-            "dom"
-        ],
-        "noImplicitAny": true,
-        "noImplicitThis": true,
-        "strictNullChecks": true,
-        "strictFunctionTypes": true,
-        "baseUrl": "../../../",
-        "typeRoots": [
-            "../../../"
-        ],
-        "paths": {
-            "node": [
-                "node/v14"
-            ]
-        },
-        "types": [],
-        "noEmit": true,
-        "forceConsistentCasingInFileNames": true
-    }
-}
