+++ types/node/v14/tsconfig.json
@@ -5,9 +5,8 @@
     ],
     "compilerOptions": {
         "module": "commonjs",
-        "target": "es6",
+        "target": "esnext",
         "lib": [
-            "es2018",
             "dom"
         ],
         "noImplicitAny": true,
